backend/requirements.txt
fastapi[standard]==0.115.12
langchain==0.3.25
sqlmodel==0.0.24
python-pptx==1.0.2
python-docx==1.1.2
langchain-google-genai==2.1.4
langchain-community==0.3.23
pdfplumber==0.11.6
fastembed==0.6.1
uvicorn==0.32.1
python-dotenv==1.0.0
aiohttp==3.10.11

# Authentication & Database
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
bcrypt==4.2.1
alembic==1.14.0
